% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotate.R
\name{annotate_ui}
\alias{annotate_ui}
\title{annotate_ui}
\usage{
annotate_ui(id)
}
\arguments{
\item{id}{A unique identifier for the Shiny namespace.}
}
\description{
annotate UI Module
}
\details{
annotate UI Module
}
