% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_merge.R
\name{data_merge_server}
\alias{data_merge_server}
\title{data_merge_server}
\usage{
data_merge_server(input, output, session)
}
\arguments{
\item{input}{description}

\item{output}{description}

\item{session}{description}
}
\description{
data_merge server Module
}
\details{
data_merge server Module
}
