% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_merge.R
\name{data_merge_ui}
\alias{data_merge_ui}
\title{data_merge_ui}
\usage{
data_merge_ui(id)
}
\arguments{
\item{id}{A unique identifier for the Shiny namespace.}
}
\description{
data_merge UI Module
}
\details{
data_merge UI Module
}
