% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_overview.R
\name{data_overview_ui}
\alias{data_overview_ui}
\title{data_overview_ui}
\usage{
data_overview_ui(id)
}
\arguments{
\item{id}{A unique identifier for the Shiny namespace.}
}
\description{
data_overview UI Module
}
\details{
data_overview UI Module
}
