% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_map.R
\name{gene_map_server}
\alias{gene_map_server}
\title{gene_map_server}
\usage{
gene_map_server(input, output, session)
}
\arguments{
\item{input}{description}

\item{output}{description}

\item{session}{description}
}
\description{
gene_map server Module
}
\details{
gene_map server Module
}
