% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_map.R
\name{gene_map_ui}
\alias{gene_map_ui}
\title{gene_map_ui}
\usage{
gene_map_ui(id)
}
\arguments{
\item{id}{A unique identifier for the Shiny namespace.}
}
\description{
gene_map UI Module
}
\details{
gene_map UI Module
}
