% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homepage.R
\name{homepage_ui}
\alias{homepage_ui}
\title{homepage_ui}
\usage{
homepage_ui(id)
}
\arguments{
\item{id}{A unique identifier for the Shiny namespace.}
}
\description{
homepage UI Module
}
\details{
homepage UI Module
}
