% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan_plot.R
\name{manhattan_plot_server}
\alias{manhattan_plot_server}
\title{manhattan_plot_server}
\arguments{
\item{input}{description}

\item{output}{description}

\item{session}{description}
}
\description{
manhattan_plot server Module
}
\details{
manhattan_plot server Module
}
