% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan_plot.R
\name{manhattan_plot_ui}
\alias{manhattan_plot_ui}
\title{manhattan_plot_ui}
\usage{
manhattan_plot_ui(id)
}
\arguments{
\item{id}{A unique identifier for the Shiny namespace.}
}
\description{
manhattan_plot UI Module
}
\details{
manhattan_plot UI Module
}
