% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNP_extraction.R
\name{SNP_extraction_server}
\alias{SNP_extraction_server}
\title{SNP_extraction_server}
\usage{
SNP_extraction_server(input, output, session)
}
\arguments{
\item{input}{description}

\item{output}{description}

\item{session}{description}
}
\description{
SNP_extraction server Module
}
\details{
SNP_extraction server Module
}
