% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNP_extraction.R
\name{SNP_extraction_ui}
\alias{SNP_extraction_ui}
\title{SNP_extraction_ui}
\usage{
SNP_extraction_ui(id)
}
\arguments{
\item{id}{A unique identifier for the Shiny namespace.}
}
\description{
SNP_extraction UI Module
}
\details{
SNP_extraction UI Module
}
